+++ spec/lightstreamer/session_spec.rb
@@ -17,41 +17,42 @@ describe Lightstreamer::Session do
       expect(stream_connection).to receive(:session_id).and_return('session')
     end
 
-    def join_processing_thread(session)
-      thread = session.instance_variable_get :@processing_thread
-      thread.join if thread
-    end
-
-    it 'connects to a stream and processes its data' do
+    it 'connects to a stream, processes some data, then disconnects' do
       expect(Lightstreamer::ControlConnection).to receive(:new)
         .with('session', 'http://test2.com')
         .and_return(control_connection)
 
+      recurring_line = "#{subscription.id},1|test"
+
+      expect(control_connection).to receive(:execute).with(:destroy) { recurring_line = nil }
+
       expect(stream_connection).to receive(:read_line).and_return("#{subscription.id},1|test")
       expect(stream_connection).to receive(:read_line).and_return('invalid data')
-      expect(stream_connection).to receive(:read_line) { Thread.exit }
-      expect(stream_connection).to receive(:disconnect)
+      allow(stream_connection).to receive(:read_line) { recurring_line }
+      expect(stream_connection).to receive(:error).and_return(nil)
 
       session.instance_variable_set :@subscriptions, [subscription]
       expect(subscription).to receive(:process_stream_data).with("#{subscription.id},1|test").and_return(true)
       expect(subscription).to receive(:process_stream_data).with('invalid data').and_return(false)
+      allow(subscription).to receive(:process_stream_data).with("#{subscription.id},1|test").and_return(true)
 
       expect do
         session.connect
-        join_processing_thread session
         session.disconnect
       end.to output("Lightstreamer: unprocessed stream data 'invalid data'\n").to_stderr
     end
 
     it 'handles when the stream connection dies' do
       expect(stream_connection).to receive(:read_line).and_return(nil)
-      expect(stream_connection).to receive(:error).and_return('test')
+      expect(stream_connection).to receive(:error).and_return(Lightstreamer::SessionEndError.new)
 
-      expect do
-        session.connect
-        join_processing_thread session
-        expect(session.connected?).to be false
-      end.to output("Lightstreamer: processing thread exiting, error: test\n").to_stderr
+      session.connect
+
+      processing_thread = session.instance_variable_get :@processing_thread
+      processing_thread.join if processing_thread
+
+      expect(session.connected?).to be false
+      expect(session.error).to be_a(Lightstreamer::SessionEndError)
     end
   end